+++ CHANGELOG.md
@@ -17,6 +17,11 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/) and this p
 - `bill`: Payment will now set issue date automatically, alongside issue time if provided.
 - `mx-cfdi-v4`: Support for "Global" B2C invoice reporting that group together B2C sales into a single document.
 - `mx-cfdi-v4`: Automatically set the `issue_time` if not already provided.
+- `tax`: `ExtensionsRequireAllOrNone` validation rule.
+
+### Removed
+
+- `tax`: `ExtensionsHas` usage should be replaced with `ExtensionsRequire` validator.
 
 ## [v0.214.1] - 2025-04-09